feat(llm): Update _validate_api_key in model suppliers (#1244)
